backend/requirements.txt
# ── NEXUS Platform — Python Dependencies ──────────────────────────────────────
# Install: pip install -r requirements.txt

# ── Web Framework ──────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
httpx==0.28.1

# ── AI / LLM ───────────────────────────────────────────────────────────────────
openai==1.57.0
langchain==0.3.10
langchain-openai==0.2.11
langchain-community==0.3.10
langgraph==0.2.56

# ── Vector Store ───────────────────────────────────────────────────────────────
chromadb==0.5.23

# ── Session Store ──────────────────────────────────────────────────────────────
redis==5.2.1

# ── Data / Validation ──────────────────────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.6.1

# ── Science / Engineering Simulation ──────────────────────────────────────────
numpy==2.1.3
scipy==1.14.1

# ── Observability ──────────────────────────────────────────────────────────────
structlog==24.4.0
opentelemetry-api==1.28.4
opentelemetry-sdk==1.28.4
opentelemetry-exporter-otlp==1.28.4
opentelemetry-instrumentation-fastapi==0.49b4

# ── Utilities ──────────────────────────────────────────────────────────────────
tenacity==9.0.0
python-dotenv==1.0.1
